+++ src/app/cart/cart.module.ts
@@ -1,22 +1,28 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {CartListComponent} from "./components/cart-list/cart-list.component";
-import {CartItemComponent} from "./components/cart-item/cart-item.component";
 import { CartComponent } from './components/cart/cart.component';
 import {SharedModule} from "../shared/shared.module";
+import {FormsModule} from "@angular/forms";
+import {MatTableModule} from "@angular/material/table";
+import {MatIconModule} from "@angular/material/icon";
+import {MatButtonModule} from "@angular/material/button";
+import {MatSortModule} from "@angular/material/sort";
 
 @NgModule({
   declarations: [
     CartListComponent,
-    CartItemComponent,
     CartComponent,
   ],
   exports: [
     CartComponent
   ],
   imports: [
-    CommonModule,
-    SharedModule
+    SharedModule,
+    MatTableModule,
+    MatIconModule,
+    MatButtonModule,
+    MatSortModule,
   ]
 })
 export class CartModule { }