+++ src/app/product/product.module.ts
@@ -3,6 +3,10 @@ import { CommonModule } from '@angular/common';
 import {ProductListComponent} from "./components/product-list-component/product-list.component";
 import {ProductItemComponent} from "./components/product-item-component/product-item.component";
 import { ProductComponent } from './components/product/product.component';
+import {MatCardModule} from "@angular/material/card";
+import {MatButtonModule} from "@angular/material/button";
+import {MatGridListModule} from "@angular/material/grid-list";
+import {FlexModule} from "@angular/flex-layout";
 
 @NgModule({
   declarations: [
@@ -14,7 +18,11 @@ import { ProductComponent } from './components/product/product.component';
     ProductComponent,
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    MatCardModule,
+    MatButtonModule,
+    MatGridListModule,
+    FlexModule
   ]
 })
 export class ProductModule { }