+++ src/app/product/services/product.repository.ts
@@ -6,11 +6,11 @@ import {Injectable} from "@angular/core";
 export class ProductRepository extends BaseRepository {
   getProducts(): Promise<ProductModel[]> {
     const productList: ProductModel[] = [
-      { id: 1, name: 'Lenovo ThinkPad x1 Carbon', description: '', category: CategoryEnum.Laptops, price: 1899, isAvailable: true, },
-      { id: 2, name: 'Lenovo ThinkPad x1 Nano', description: '', category: CategoryEnum.Laptops, price: 2099, isAvailable: false, },
-      { id: 3, name: 'Lenovo ThinkPad x1 Extreme', description: '', category: CategoryEnum.Laptops, price: 2299, isAvailable: true, },
-      { id: 4, name: 'MacBook Air M1 2020', description: '', category: CategoryEnum.Laptops, price: 999, isAvailable: false, },
-      { id: 5, name: 'MacBook Pro M1 2020', description: '', category: CategoryEnum.Laptops, price: 1199, isAvailable: true, },
+      { id: 1, name: 'Lenovo ThinkPad x1 Carbon', description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec rutrum.', category: CategoryEnum.Laptops, price: 1899.99, isAvailable: true, },
+      { id: 2, name: 'Lenovo ThinkPad x1 Nano', description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec rutrum.', category: CategoryEnum.Laptops, price: 2099, isAvailable: false, },
+      { id: 3, name: 'Lenovo ThinkPad x1 Extreme', description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec rutrum.', category: CategoryEnum.Laptops, price: 2299, isAvailable: true, },
+      { id: 4, name: 'MacBook Air M1 2020', description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec rutrum.', category: CategoryEnum.Laptops, price: 999.99, isAvailable: false, },
+      { id: 5, name: 'MacBook Pro M1 2020', description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec rutrum.', category: CategoryEnum.Laptops, price: 1199, isAvailable: true, },
     ];
 
     return Promise.resolve(productList);