+++ src/app/shared/components/toolbar/toolbar.component.ts
@@ -1,0 +1,19 @@
+import {AfterViewInit, Component, ElementRef, OnInit, ViewChild} from '@angular/core';
+
+@Component({
+  selector: 'app-toolbar',
+  templateUrl: './toolbar.component.html',
+  styleUrls: ['./toolbar.component.scss']
+})
+export class ToolbarComponent implements OnInit, AfterViewInit {
+  @ViewChild('appTitle') title!: ElementRef<HTMLHeadingElement>;
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+  ngAfterViewInit(): void {
+    this.title.nativeElement.innerText = 'Test App';
+  }
+}