+++ src/app/shared/models/dropdown-select-item.model.ts
@@ -1,0 +1,8 @@
+export class DropdownSelectItemModel<T> {
+  constructor(
+    public label: string,
+    public value: T,
+    public isDefault: boolean,
+  ) {
+  }
+}